+++ src/main/java/me/reimnop/d4f/customevents/actions/Action.java
@@ -4,5 +4,5 @@ import com.google.gson.JsonElement;
 import me.reimnop.d4f.customevents.ActionContext;
 
 public interface Action {
-    void runAction(JsonElement value, ActionContext context);
+    void runAction(JsonElement value, ActionContext context, int num);
 }