+++ src/main/java/me/reimnop/d4f/customevents/constraints/LinkedAccountNickConstraintProcessor.java
@@ -42,7 +42,7 @@ public class LinkedAccountNickConstraintProcessor implements ConstraintProcessor
             return false;
         }
 
-        Member member = Discord4Fabric.DISCORD.getMember(discordProfile.get().id);
+        Member member = Discord4Fabric.DISCORD.getMember(discordProfile.get().id, 0);
         if (member == null) {
             return false;
         }