+++ src/main/java/me/reimnop/d4f/listeners/DiscordCommandProcessor.java
@@ -4,8 +4,8 @@ import me.drex.vanish.api.VanishAPI;
 import me.reimnop.d4f.Config;
 import me.reimnop.d4f.Discord;
 import me.reimnop.d4f.Discord4Fabric;
+import me.reimnop.d4f.Storage;
 import me.reimnop.d4f.events.DiscordMessageReceivedCallback;
-import me.reimnop.d4f.utils.Compatibility;
 import me.reimnop.d4f.utils.Utils;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.server.MinecraftServer;
@@ -25,8 +25,8 @@ public final class DiscordCommandProcessor {
     }
 
     private static final Map<String, DiscordCommandHandler> commandHandlers = Map.of(
-            "ping", server -> Discord4Fabric.DISCORD.sendPlainMessage("Pong! :ping_pong:"),
-            "tps", server -> Discord4Fabric.DISCORD.sendPlainMessage("Server TPS: " + Utils.getTpsAsString()),
+            "ping", server -> Discord4Fabric.DISCORD.sendPlainMessage("Pong! :ping_pong:",0),
+            "tps", server -> Discord4Fabric.DISCORD.sendPlainMessage("Server TPS: " + Utils.getTpsAsString(),0),
             "playerlist", server -> {
                 int maxPlayers = server.getMaxPlayerCount();
                 String[] playerNames = server.getPlayerNames();
@@ -63,15 +63,15 @@ public final class DiscordCommandProcessor {
 
                 }
 
-                Discord4Fabric.DISCORD.sendPlainMessage(stringBuilder.toString());
+                Discord4Fabric.DISCORD.sendPlainMessage(stringBuilder.toString(),0);
             }
     );
 
     private static final String prefix = "!";
 
-    public static void init(Config config) {
+    public static void init(Config config, Storage storage, int num) {
         DiscordMessageReceivedCallback.EVENT.register((user, message) -> {
-            if (message.getChannel().getIdLong() != config.channelId) {
+            if (message.getChannel().getIdLong() != storage.channelId[num]) {
                 return;
             }
 
@@ -102,17 +102,13 @@ public final class DiscordCommandProcessor {
      */
     private static String[] filterDrexHDVanishedPlayers(MinecraftServer server, String[] unfilteredNames) {
 
-        if (!Compatibility.isVanishModLoaded()) {
-            return unfilteredNames;
-        }
-
         return Arrays.stream(unfilteredNames)
                 .filter(
                         name -> {
                             PlayerManager manager = server.getPlayerManager();
                             ServerPlayerEntity player = manager.getPlayer(name);
                             // player should never be null as the names are assumed to all be valid
-                            return player != null && Compatibility.isPlayerVanished(player);
+                            return player != null;
                         }
                 )
                 .toArray(String[]::new);